feat: introduce AutomatonRuntime trait for improved state management
- Added the `AutomatonRuntime` trait to provide a stable interface for reading and writing the automaton's runtime state, decoupling helper modules from the `GraphAutomaton` type.
- Implemented the `AutomatonRuntime` trait for `GraphAutomaton`, allowing access to the underlying graph, tick count, tick history, and configuration.
- Updated various modules to utilize the new trait, enhancing flexibility in automaton state handling and improving code maintainability.

+++ crates/vibe-graph-automaton/src/description/generator.rs
@@ -672,7 +672,7 @@ mod tests {
                 },
             });
             edges.push(GraphEdge {
-                id: EdgeId(i as u64),
+                id: EdgeId(i),
                 from: NodeId(i),
                 to: NodeId(100),
                 relationship: "imports".to_string(),
